requirements.txt
requests
Django==1.10.3
django-bower==5.2.0
pymongo==3.3.1
six==1.10.0
navitia_client
pandas
numpy
gevent
ipdb
djangorestframework
boto3
